Better Fix for flask limiter 4.0
