+++ cps/kobo_auth.py
@@ -62,12 +62,13 @@ particular calls to non-Kobo specific endpoints such as the CalibreWeb book down
 from binascii import hexlify
 from datetime import datetime
 from os import urandom
+from functools import wraps
 
-from flask import g, Blueprint, request
-from .cw_login import current_user
+from flask import g, Blueprint, abort, request
+from .cw_login import login_user, current_user
 from flask_babel import gettext as _
 
-from . import logger, config, calibre_db, db, helper, ub, lm
+from . import logger, config, calibre_db, db, helper, ub, lm, limiter
 from .render_template import render_title_template
 from .usermanagement import user_login_required
 
@@ -133,6 +134,13 @@ def disable_failed_auth_redirect_for_blueprint(bp):
     lm.blueprint_login_views[bp.name] = None
 
 
+def get_auth_token():
+    if "auth_token" in g:
+        return g.get("auth_token")
+    else:
+        return None
+
+
 def register_url_value_preprocessor(kobo):
     @kobo.url_value_preprocessor
     # pylint: disable=unused-variable
@@ -140,3 +148,21 @@ def register_url_value_preprocessor(kobo):
         g.auth_token = values.pop("auth_token")
 
 
+def requires_kobo_auth(f):
+    @wraps(f)
+    def inner(*args, **kwargs):
+        auth_token = get_auth_token()
+        if auth_token is not None:
+            user = (
+                ub.session.query(ub.User)
+                .join(ub.RemoteAuthToken)
+                .filter(ub.RemoteAuthToken.auth_token == auth_token).filter(ub.RemoteAuthToken.token_type==1)
+                .first()
+            )
+            if user is not None:
+                login_user(user)
+                [limiter.limiter.clear(limit.limit, *limit.request_args) for limit in limiter.current_limits]
+                return f(*args, **kwargs)
+        log.debug("Received Kobo request without a recognizable auth token.")
+        return abort(401)
+    return inner
